refactor: split prompt-lane assembly and the context/cost report out of main.rs
Two topics that had nothing to do with the REPL loop they were living next to,
and — measured before the cut — nothing to do with the rest of main.rs either:
the region needed ZERO symbols from it.

  agent::prompt_lanes   how the system prompt is ordered (frozen core, persona,
                        skills, then the per-turn dynamic lane) and what a turn
                        folds into the message it sends. The ordering is a
                        prefix-cache contract, which the module doc now states
                        instead of leaving it spread across call sites.
  ui::context_report    /tokens, /context, /cost — read-only reporting, plus the
                        window-size heuristic and why it is deliberately not a cap.

`context_breakdown`'s tests move with their subject; their `#[path]` is now
relative to src/ui/. One clippy suggestion ("identical blocks") travels unchanged
for the same reason as the last move: not a behaviour change to smuggle in here.

main.rs: 7,546 -> 6,960
