+++ src/main.rs
@@ -27,6 +27,7 @@ mod cli_args;
 // call sites in THIS file referring to the modules by their short names (no behavior
 // change) — every other file already uses the new `crate::<group>::<mod>` paths.
 use crate::agent::app_catalog;
+use crate::agent::prompt_lanes::*;
 use crate::core::session_store::*;
 use crate::core::{cli_config, config, types};
 use crate::features::slash_handlers::{
@@ -37,6 +38,7 @@ use crate::hostbot::platforms::telegram;
 use crate::llm::client;
 use crate::persona::soul;
 use crate::skills::{self as skill, registry as skill_registry};
+use crate::ui::context_report::*;
 use crate::ui::menus::{apps_menu, persona_self_view_n, run_discord, run_telegram};
 use crate::ui::{config_ui, icons, image_input, splash, theme, tui};
 
@@ -4136,594 +4138,6 @@ async fn run_persona_reflection(
         );
     }
 }
-
-/// Build prompt lanes around a caller-selected frozen core. Keeping the lifecycle choice OUT of this
-/// helper makes every call site say whether it is opening a fresh conversation (refresh/adopt) or
-/// merely rewriting lanes inside the current one (read the already-adopted bytes).
-fn system_prompt_bundle_with_core(model: &str, frozen: &str) -> agent::PromptBundle {
-    system_prompt_bundle_in(model, frozen, None)
-}
-
-/// As `system_prompt_bundle_with_core`, but stating an EXPLICIT working directory in the prompt.
-/// The hostbot daemon passes its lane's cwd: the prompt tells the model where it is working, and
-/// under several concurrent bots the process cwd is not that answer for any of them.
-fn system_prompt_bundle_in(
-    model: &str,
-    frozen: &str,
-    root: Option<&std::path::Path>,
-) -> agent::PromptBundle {
-    let cwd = root
-        .map(|p| p.display().to_string())
-        .or_else(|| {
-            std::env::current_dir()
-                .ok()
-                .map(|p| p.display().to_string())
-        })
-        .unwrap_or_else(|| ".".to_string());
-    let date = chrono::Local::now().format("%Y-%m-%d").to_string();
-    let mut bundle = agent::build_top_level_system_prompt_bundle(
-        &cwd,
-        std::env::consts::OS,
-        &date,
-        model,
-        Some(frozen),
-    );
-    // L2 session working memory (temporary, budget-capped). Empty → no tag (zero cost).
-    let sess_budget = memory::settings().session_mem_max_tokens;
-    if let Some(block) = memory::session_mem::process_prompt_block(sess_budget) {
-        bundle.dynamic.push('\n');
-        bundle.dynamic.push_str(&block);
-        bundle.dynamic.push('\n');
-    }
-    bundle
-}
-
-/// A TRUE conversation boundary: promote pending memory, rebuild from the current store, and adopt
-/// the result before constructing a fresh prompt prefix. Startup, `/clear`, `/handoff`, session load,
-/// and one-shot/captured runs are the only callers that should use this path.
-fn refreshed_system_prompt_bundle(model: &str) -> agent::PromptBundle {
-    let frozen = memory::refresh_frozen_core();
-    system_prompt_bundle_with_core(model, &frozen)
-}
-
-/// Same-conversation lane rewrite: reuse the already-adopted core byte-for-byte. A retrieval,
-/// reinforcement, or memory write during this conversation may stage `core.next`, but must not
-/// mutate the cached prefix or promote it before the next conversation boundary.
-fn active_system_prompt_bundle(model: &str) -> agent::PromptBundle {
-    let frozen = memory::active_frozen_core();
-    system_prompt_bundle_with_core(model, &frozen)
-}
-
-/// The two bundle builders above, parameterized by the hostbot lane's working directory and persona.
-/// Both are needed together because the persona override must be armed for the (synchronous) prompt
-/// assembly and released immediately after — see `persona::with_override`.
-fn hostbot_prompt_bundle(
-    model: &str,
-    root: &std::path::Path,
-    persona: Option<String>,
-    boundary: bool,
-) -> agent::PromptBundle {
-    persona::with_override(persona, || {
-        let frozen = if boundary {
-            memory::refresh_frozen_core()
-        } else {
-            memory::active_frozen_core()
-        };
-        system_prompt_bundle_in(model, &frozen, Some(root))
-    })
-}
-
-/// Seed both system lanes for a brand-new conversation.
-fn seed_prompt_lanes(history: &mut Vec<Message>, model: &str) {
-    history.clear();
-    let bundle = refreshed_system_prompt_bundle(model);
-    history.push(Message::system(bundle.stable));
-    if !bundle.dynamic.trim().is_empty() {
-        history.push(Message::system(bundle.dynamic));
-    }
-}
-
-/// Replace a persisted zero/one-system legacy prefix with the current two-lane prompt.
-/// Histories already carrying both lanes are left byte-identical.
-pub(crate) fn migrate_legacy_prompt_lanes(history: &mut Vec<Message>, model: &str) {
-    let lead = agent::compact::leading_system_count(history);
-    if lead >= 2 {
-        return;
-    }
-    let tail = history.get(lead..).unwrap_or_default().to_vec();
-    seed_prompt_lanes(history, model);
-    history.extend(tail);
-}
-
-/// Per-turn budget (tokens, chars/4 estimate) for the `/init` codebase-retrieval block folded into
-/// the CURRENT user turn (see [`fold_retrieval_into_query`]). Small enough to stay well under the
-/// frozen-core/session budgets but big enough for ~5-8 chunks with attribution.
-const CODEBASE_RETRIEVAL_BUDGET_TOKENS: usize = 1500;
-
-/// Fresh user-turn boundary: refresh only the dynamic lane, preserving stable index 0 byte-for-byte.
-fn refresh_dynamic_prompt_lane(history: &mut Vec<Message>, model: &str) {
-    migrate_legacy_prompt_lanes(history, model);
-    let dynamic = active_system_prompt_bundle(model).dynamic;
-    let lead = agent::compact::leading_system_count(history);
-    if dynamic.trim().is_empty() {
-        if lead > 1 {
-            history.remove(1);
-        }
-    } else if lead > 1 {
-        history[1] = Message::system(dynamic);
-    } else {
-        history.insert(1, Message::system(dynamic));
-    }
-}
-
-/// Rewrite BOTH system lanes in place, preserving every non-system message.
-///
-/// For settings changes that alter the STABLE lane — `/model` and `/config` both do, since the model
-/// name, prompt tier and `<project_context>` live at index 0 — but must NOT end the conversation.
-/// `rebuild_system` cannot serve here: it calls `seed_prompt_lanes`, which starts with
-/// `history.clear()`, so using it for a settings change silently threw away the whole chat (the user
-/// went to `/config` to retune the context and came back to an empty thread).
-///
-/// Session working memory is deliberately KEPT: this is the same conversation, so its scratch notes
-/// are still valid. That is the other half of why `/config` must not route through `rebuild_system`,
-/// which drops them as part of starting a new thread.
-/// Splice a caller-selected prompt bundle over the leading prompt lanes while preserving every
-/// conversation message (including a handoff seed, which is a third system message but NOT part of
-/// the two-lane prefix).
-fn splice_prompt_lanes(history: &mut Vec<Message>, bundle: agent::PromptBundle) {
-    let lead = agent::compact::leading_system_count(history);
-    let mut lanes = vec![Message::system(bundle.stable)];
-    if !bundle.dynamic.trim().is_empty() {
-        lanes.push(Message::system(bundle.dynamic));
-    }
-    history.splice(0..lead, lanes);
-}
-
-/// Same-conversation rewrite (`/config`, `/model`, persona change): keep the active core stable.
-pub(crate) fn refresh_prompt_lanes_in_place(history: &mut Vec<Message>, model: &str) {
-    splice_prompt_lanes(history, active_system_prompt_bundle(model));
-}
-
-/// Thread switch (`/resume`, session/time-machine restore): refresh/adopt memory for the new
-/// conversation before rebuilding the current-project prompt lanes around its saved transcript.
-pub(crate) fn refresh_prompt_lanes_for_thread_switch(history: &mut Vec<Message>, model: &str) {
-    splice_prompt_lanes(history, refreshed_system_prompt_bundle(model));
-}
-
-/// Automatic codebase RAG, folded into the CURRENT user turn (NOT the dynamic system lane).
-///
-/// When `/init` has built an index, the top-ranked chunks (path + line range + real content,
-/// source-attributed) are prepended to the user's message so the model sees relevant code before it
-/// even calls a tool. Placing it on the user turn — the volatile, already-uncached message — keeps
-/// index 1 (the dynamic system lane) byte-stable, so the provider's prefix cache still covers the
-/// whole transcript tail up to the last stable turn. Folding into the dynamic lane instead would
-/// vary index 1 every turn and force the entire transcript after it to re-bill uncached (the
-/// Anthropic prefix-cache breakpoint sits on the last stable assistant/tool message).
-///
-/// Returns the message content to send. The caller keeps the ORIGINAL `query` for checkpoint /
-/// display / persisted history — only the sent content carries the (ephemeral, per-turn) block.
-/// No-op passthrough when there is no index / no query terms / nothing clears the relevance gate.
-fn fold_retrieval_into_query(query: &str) -> String {
-    if query.trim().is_empty() {
-        return query.to_string();
-    }
-    // Kick a background drift check: if source files changed since the last /init, an incremental
-    // rebuild runs off-turn so the NEXT turn sees fresh context. Never blocks this turn (#17).
-    crate::agent::codebase::ensure_fresh();
-    match crate::agent::codebase::retrieval_block(query, CODEBASE_RETRIEVAL_BUDGET_TOKENS) {
-        Some(block) => format!("{block}\n\n{query}"),
-        None => query.to_string(),
-    }
-}
-
-/// Per-turn budget (tokens) for the memory recall block folded into the CURRENT user turn.
-/// Deliberately an order of magnitude under the codebase budget: this carries a handful of
-/// one-line facts, not source, and it is spent on every gated turn.
-const MEMORY_RECALL_BUDGET_TOKENS: usize = 300;
-
-/// Fold BOTH per-turn context blocks into the sent content: memory recall, then codebase RAG.
-///
-/// Same discipline as [`fold_retrieval_into_query`] and for the same reason — the blocks ride on the
-/// **user turn**, which is already uncached, so system lanes 0/1 stay byte-stable and the provider's
-/// prefix cache keeps covering the transcript tail (invariant I1).
-///
-/// Memory goes FIRST so the standing facts ("reply in Vietnamese", "windows-sys is pinned") are read
-/// before the code they qualify. The recall block also seats its handle→id pairs in the pending
-/// ledger, which is what lets a later `used` report confirm only facts that were actually shown.
-///
-/// `query` itself is never modified: the caller keeps it for checkpoint / display / persisted
-/// history, so the durable transcript holds the user's real words, not our scaffolding.
-fn fold_context_into_query(query: &str) -> String {
-    // The turn counter lives here because this is the one point BOTH REPL loops pass through exactly
-    // once per user message. Counting inside the agent loop would count iterations, and metric 1's
-    // denominator ("live facts per turn") has to mean turns the user drove.
-    memory::stats::note_turn();
-    let mut out = fold_retrieval_into_query(query);
-    // Skills that actually fit THIS question, gated on the same coverage threshold as recall. The
-    // always-on `<skills>` index names every applicable procedure regardless of the request; this
-    // block is what makes the fitting ones salient without spending the system lane's byte-stable
-    // budget on the ones that don't. Folded ABOVE the code but BELOW the facts, matching the
-    // "standing truth → how-to → source" reading order.
-    if let Some(block) = skills::turn_block(query, skills::SKILL_TURN_BUDGET_TOKENS) {
-        out = format!("{block}\n\n{out}");
-    }
-    if let Some((block, pairs)) = memory::recall_block(query, MEMORY_RECALL_BUDGET_TOKENS) {
-        memory::pending::open_turn(pairs);
-        out = format!("{block}\n\n{out}");
-    }
-    out
-}
-
-/// Drop per-turn context blocks (memory recall, gated skills) from user turns already in `history`.
-///
-/// Each block was true for the turn it rode in on. Left in place they accumulate — ten turns of
-/// standing facts re-stated ten times — and, worse, an older block can contradict a newer one with
-/// nothing in the transcript marking which came later, so the model has to guess.
-///
-/// Called only from [`maybe_auto_compact`], at the moment the prefix cache is being invalidated
-/// anyway: rewriting a user turn at any other time would break cache coverage for the whole tail,
-/// costing more than the tokens it saves.
-///
-/// Matches on [`memory::RECALL_MARKER`] / [`skills::SKILL_MARKER`] at the start of the content and
-/// cuts through the first blank line. Anything the user actually typed survives, including a message
-/// that merely mentions the phrase — a marker has to be at position 0, which only our own folding
-/// produces. Both are peeled in a loop because a turn carries them stacked (recall, then skills):
-/// removing the outer one promotes the inner one to position 0.
-fn strip_recall_blocks(history: &mut [Message]) {
-    for m in history.iter_mut() {
-        if m.role != "user" {
-            continue;
-        }
-        let Some(content) = m.content.as_deref() else {
-            continue;
-        };
-        let mut cur = content;
-        loop {
-            let next = strip_skill_prefix(memory::strip_recall_prefix(cur));
-            if next.len() == cur.len() {
-                break;
-            }
-            cur = next;
-        }
-        if cur.len() != content.len() {
-            m.content = Some(cur.to_string());
-        }
-    }
-}
-
-/// Peel one leading gated-skill block, mirroring [`memory::strip_recall_prefix`].
-fn strip_skill_prefix(content: &str) -> &str {
-    if !content.starts_with(skills::SKILL_MARKER) {
-        return content;
-    }
-    match content.split_once("\n\n") {
-        Some((_, rest)) => rest,
-        None => content,
-    }
-}
-
-/// Everything a THREAD SWITCH must reset besides history itself: session scratch memory, todos,
-/// the cost tally, destructive-op session grants, and browser page @refs. `/clear`, `/handoff`,
-/// `/resume`, `/sessions` restore and `/recover` all route here so a fresh or restored thread
-/// never inherits the previous one's state (the classic leak: a restored conversation still
-/// "allowed" the old thread's destructive ops and showed its cost).
-pub(crate) fn reset_per_session_state() {
-    memory::session_mem::clear_process_session_mem();
-    // The new transcript never contained the old recall block, so its handles now point at facts
-    // the model cannot see — and a stale `last_ids` would suppress the first block of the new
-    // thread as a "duplicate" of one that is no longer in context.
-    memory::pending::clear();
-    crate::agent::todo::clear();
-    client::cost_meter().reset();
-    tui::reset_session_allow();
-    #[cfg(feature = "browser")]
-    crate::agent::browser::release_active();
-}
-
-/// Reset the conversation to just the system prompt (fresh session / model change). Rebuilds the
-/// frozen core from the current memory store so newly added `type=user` facts / STYLE are injected.
-/// Drops session working memory — a new thread does not inherit this session's scratch notes.
-pub(crate) fn rebuild_system(history: &mut Vec<Message>, model: &str) {
-    memory::session_mem::clear_process_session_mem();
-    seed_prompt_lanes(history, model);
-}
-
-/// Replace the system lanes in place WITHOUT clearing the conversation — used when switching
-/// persona mid-chat so the new character applies but the history is preserved.
-pub(crate) fn update_system_prompt(history: &mut Vec<Message>, model: &str) {
-    refresh_dynamic_prompt_lane(history, model);
-}
-
-/// Approximate context window (tokens) for a model, by name pattern. A rough heuristic for the
-/// `% context` HUD only — not a hard cap (the upstream enforces the real limit). Defaults to 128K.
-fn ctx_window_for(model: &str) -> usize {
-    let m = model.to_ascii_lowercase();
-    if m.contains("1m") {
-        1_000_000 // explicit 1M-context variants (e.g. opus-4-8-1m-thinking) — checked before the family heuristics
-    } else if m.contains("gemini") {
-        1_000_000
-    } else if m.contains("claude")
-        || m.contains("opus")
-        || m.contains("sonnet")
-        || m.contains("haiku")
-    {
-        200_000
-    } else if m.contains("gpt-4.1") || m.contains("o3") || m.contains("o4") {
-        1_000_000
-    } else if m.contains("deepseek") {
-        64_000
-    } else {
-        128_000 // gpt-4o family + safe default
-    }
-}
-
-/// A 10-cell context-fill bar, coloured by pressure using the semantic palette (P-ctx4): OK below
-/// 50%, WARN gold from 50%, ERR salmon from 80% — the same green/gold/salmon meanings the rest of
-/// the UI uses, instead of bespoke 256-colour indices.
-fn ctx_bar(pct: f64) -> String {
-    const CELLS: usize = 10;
-    let filled = ((pct / 100.0) * CELLS as f64)
-        .round()
-        .clamp(0.0, CELLS as f64) as usize;
-    let bar = format!("{}{}", "█".repeat(filled), "░".repeat(CELLS - filled));
-    let color: u8 = if pct >= 80.0 {
-        theme::ERR
-    } else if pct >= 50.0 {
-        theme::WARN
-    } else {
-        theme::OK
-    };
-    style(bar).color256(color).to_string()
-}
-
-/// The effective window from an explicit/configured value (when present) over the name heuristic.
-/// Returns `(tokens, was_configured)`. Pure — callers pass the value (lets the wizard compute it
-/// against unsaved in-memory config).
-pub(crate) fn effective_ctx_window(model: &str, configured: Option<usize>) -> (usize, bool) {
-    match configured {
-        Some(w) if w > 0 => (w, true),
-        _ => (ctx_window_for(model), false),
-    }
-}
-
-/// The effective context window for `model`: a provider-reported/manually-set value in config (when
-/// it matches the active model) wins over the name heuristic. Returns `(tokens, was_configured)`.
-fn resolve_ctx_window(model: &str) -> (usize, bool) {
-    let cfg = cli_config::load();
-    let configured = cfg
-        .model_context_window
-        .filter(|_| cfg.model.as_deref() == Some(model));
-    effective_ctx_window(model, configured)
-}
-
-/// Rough session size in tokens — shared by the HUD + auto-compact. Delegates to the agent
-/// estimator (content + tool-call payloads + envelopes) plus the tool-schema overhead the loop
-/// last published, so the HUD and the mid-loop guards agree on request size.
-fn session_tokens(history: &[Message]) -> usize {
-    history
-        .iter()
-        .map(agent::estimate_message_tokens)
-        .sum::<usize>()
-        + agent::schema_overhead_tokens()
-}
-
-/// Compact a token count for display: `12.4K` / `300`.
-fn fmt_k(n: usize) -> String {
-    if n >= 1000 {
-        format!("{:.1}K", n as f64 / 1000.0)
-    } else {
-        n.to_string()
-    }
-}
-
-/// `/cost` — session token accounting + (when rates are set) an estimated $ cost. Honest by design:
-/// shows REAL provider-reported tokens when the endpoint sends `usage`, else the chars/4 context
-/// estimate clearly labelled — and never invents a price or a credit balance.
-pub(crate) fn print_cost(history: &[Message], model: &str) {
-    let (p, c, calls) = client::cost_meter().snapshot();
-    let cfg = cli_config::load();
-    if calls > 0 {
-        let total = p + c;
-        let mut line = format!(
-            "{}  {} in + {} out = {} tok  ({} call{} reported usage)",
-            style("💰 session usage").color256(splash::ACCENT).bold(),
-            fmt_k(p as usize),
-            fmt_k(c as usize),
-            fmt_k(total as usize),
-            calls,
-            if calls == 1 { "" } else { "s" },
-        );
-        match (cfg.price_in, cfg.price_out) {
-            (Some(pin), Some(pout)) => {
-                let cost = p as f64 / 1_000_000.0 * pin + c as f64 / 1_000_000.0 * pout;
-                line.push_str(&format!(
-                    "  ·  {}",
-                    style(format!("est ${cost:.4} (@ ${pin}/${pout} per 1M in/out)")).color256(splash::ACCENT)
-                ));
-            }
-            _ => line.push_str(&format!(
-                "  ·  {}",
-                style("set rates for a $ estimate: aizen config set --price-in <$/1M> --price-out <$/1M>").dim()
-            )),
-        }
-        // Prompt-cache payoff (only when the provider reported cache reads → confirms caching works).
-        let cached = client::cost_meter().cache_read();
-        if cached > 0 {
-            line.push_str(&format!(
-                "  ·  {}",
-                style(format!("{} cached @ ~0.1× in", fmt_k(cached as usize))).color256(theme::OK)
-            ));
-        }
-        tui::emit_line(&line);
-    } else {
-        // No real usage from the provider → fall back to the context-size estimate (not a $ figure).
-        let est = session_tokens(history);
-        let (window, _) = resolve_ctx_window(model);
-        tui::emit_line(&format!(
-            "{}  ~{} tok in context · window {} {}",
-            style("📊 estimated").color256(splash::ACCENT).bold(),
-            fmt_k(est),
-            fmt_k(window),
-            style("(chars/4 — the provider didn't report token usage, so no per-call $ to show)")
-                .dim()
-        ));
-    }
-}
-
-/// Decompose the live system prompt into its named blocks by XML tag, returning (label, char count)
-/// for the leftover base instructions plus every block actually present. Pure (byte-index scan over
-/// ASCII tags) so it's unit-testable; char counts ÷4 ≈ tokens, the same basis the HUD estimator uses.
-fn system_block_chars(system: &str) -> Vec<(&'static str, usize)> {
-    // (display label, tag) in build order — an absent block contributes nothing.
-    const BLOCKS: &[(&str, &str)] = &[
-        ("environment", "environment"),
-        ("agent identity", "agent_identity"),
-        ("persona", "persona"),
-        ("persona memory", "self"),
-        ("user memory", "user_memory"),
-        ("skills index", "skills"),
-        ("project context", "project_context"),
-        ("agents index", "agents"),
-    ];
-    let mut rows = Vec::new();
-    let mut tagged = 0usize;
-    for (label, tag) in BLOCKS {
-        let open = format!("<{tag}>");
-        let close = format!("</{tag}>");
-        if let (Some(s), Some(e)) = (system.find(&open), system.find(&close)) {
-            if e >= s {
-                // Tags are ASCII, so byte slicing lands on char boundaries.
-                let c = system[s..e + close.len()].chars().count();
-                tagged += c;
-                rows.push((*label, c));
-            }
-        }
-    }
-    let base = system.chars().count().saturating_sub(tagged);
-    let mut out = vec![("base instructions", base)];
-    out.extend(rows);
-    out
-}
-
-/// Render the `/compact` result as a small tree: a headline with the token delta, then one `└` leaf
-/// per file the collapsed turns had referenced and one for the skills they had loaded. The leaves are
-/// what makes compaction feel non-lossy — the dense summary note is invisible, but this shows at a
-/// glance the concrete context (which files, which skills) those turns carried, harvested by
-/// [`agent::compact::context_touchpoints`] BEFORE the collapse.
-pub(crate) fn print_compact_summary(before: usize, after: usize, tp: &agent::compact::Touchpoints) {
-    let saved = before.saturating_sub(after);
-    tui::emit_line(&format!(
-        "{}  {} → {} tok{}",
-        style("✳ Compacted").color256(splash::ACCENT).bold(),
-        style(format!("~{}", fmt_k(before))).dim(),
-        style(format!("~{}", fmt_k(after))).color256(splash::ACCENT),
-        if saved > 0 {
-            style(format!("  · freed ~{}", fmt_k(saved)))
-                .dim()
-                .to_string()
-        } else {
-            String::new()
-        },
-    ));
-    let leaf = style("  └").color256(theme::FAINT).to_string();
-    for f in &tp.files {
-        tui::emit_line(&format!(
-            "{leaf} {} {}",
-            style("Referenced file").dim(),
-            style(f).color256(theme::ACCENT_DIM)
-        ));
-    }
-    if !tp.skills.is_empty() {
-        tui::emit_line(&format!(
-            "{leaf} {} ({})",
-            style("Skills restored").dim(),
-            style(tp.skills.join(", ")).color256(theme::ACCENT_DIM),
-        ));
-    }
-    if tp.files.is_empty() && tp.skills.is_empty() {
-        tui::emit_line(&format!(
-            "{leaf} {}",
-            style("no files or skills to carry forward").dim()
-        ));
-    }
-}
-
-/// `/context` — where the tokens are going right now: the system prompt split into its blocks, the
-/// tool-schema overhead (rides every request, lives in no message), and the conversation split by
-/// role. Estimated (chars/4) — the same honest basis the HUD + auto-compact use; `/cost` shows the
-/// provider's REAL billed count when the endpoint reports usage.
-pub(crate) fn print_context(history: &[Message], model: &str) {
-    let (window, auto) = resolve_ctx_window(model);
-    let total = session_tokens(history);
-    let pct = (total as f64 / window as f64 * 100.0).min(100.0);
-
-    let system = history
-        .first()
-        .filter(|m| m.role == "system")
-        .and_then(|m| m.content.as_deref())
-        .unwrap_or("");
-    let sys_blocks = system_block_chars(system);
-    let sys_tok: usize = sys_blocks.iter().map(|(_, c)| c / 4).sum();
-    let schemas = agent::schema_overhead_tokens();
-
-    // Everything after the system message, bucketed by role.
-    let (mut user_tok, mut asst_tok, mut tool_tok) = (0usize, 0usize, 0usize);
-    for m in history.iter().skip(1) {
-        let t = agent::estimate_message_tokens(m);
-        match m.role.as_str() {
-            "assistant" => asst_tok += t,
-            "tool" => tool_tok += t,
-            _ => user_tok += t, // user turns + any stray system nudges
-        }
-    }
-    let convo = user_tok + asst_tok + tool_tok;
-
-    // One aligned row: label left-padded to a column, "~X.XK tok" right; sub-rows dimmed + indented.
-    fn line(label: &str, tok: usize, depth: usize, dim: bool) -> String {
-        let name = format!("{}{}", "  ".repeat(depth), label);
-        let s = format!("{name:<26} {:>10}", format!("~{} tok", fmt_k(tok)));
-        if dim {
-            style(s).dim().to_string()
-        } else {
-            s
-        }
-    }
-
-    tui::emit_line(&format!(
-        "{}  {model} · window {}{}",
-        style("📊 context breakdown")
-            .color256(splash::ACCENT)
-            .bold(),
-        fmt_k(window),
-        if auto { "" } else { " (est)" },
-    ));
-    tui::emit_line(&line("system prompt", sys_tok, 0, false));
-    for (label, c) in &sys_blocks {
-        if c / 4 > 0 {
-            tui::emit_line(&line(label, c / 4, 1, true));
-        }
-    }
-    tui::emit_line(&line("tool schemas", schemas, 0, false));
-    tui::emit_line(&line("conversation", convo, 0, false));
-    if convo > 0 {
-        tui::emit_line(&line("user turns", user_tok, 1, true));
-        tui::emit_line(&line("assistant turns", asst_tok, 1, true));
-        tui::emit_line(&line("tool results", tool_tok, 1, true));
-    }
-    let bar = format!("{} {}", ctx_bar(pct), style(format!("{pct:.0}%")).dim());
-    tui::emit_line(&format!(
-        "{}  {} {bar}",
-        style(format!("{:<26}", "total"))
-            .color256(splash::ACCENT)
-            .bold(),
-        style(format!("~{} / {} tok", fmt_k(total), fmt_k(window))).color256(splash::ACCENT),
-    ));
-}
-
-#[cfg(test)]
-#[path = "tests/context_breakdown.rs"]
-mod context_breakdown_tests;
-
 /// Tell the user, unmistakably, when a turn ended for a reason that is NOT success.
 ///
 /// The agent loop can return with the work unfinished or the tree broken, and in those cases the
