+++ src/ui/mod.rs
@@ -5,6 +5,7 @@
 pub mod cards;
 pub mod channel_markdown;
 pub mod config_ui;
+pub mod context_report;
 pub mod icons;
 pub mod image_input;
 pub mod links;
